✨ Feat: 운동 정보 게시판 CRUD
- 임시 커밋

+++ src/main/java/com/spring/myapp/informationBoard/model/Likes.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class Likes {
-	private Integer infoId;
-	private Integer userId;
+	private Long infoId;
+	private Long userId;
 	private LocalDateTime createdAt;
 }
